Fixed NoSuchMethod
